feat #22 Unit tests
-correction brought by tests,
-cors for the front client

+++ CatMash.DAL/Sql/CreateTables.sql
@@ -21,7 +21,7 @@ CREATE TABLE dbo.Cats
 	IsTopOne BIT NOT NULL,
 	IsAlone BIT NOT NULL,
 	Rating FLOAT NOT NULL,
-	ViewsNumber FLOAT NOT NULL,
+	ViewsNumber INT NOT NULL,
 	ProbabilityWeight FLOAT NOT NULL,
 	INDEX IX_IsAlone NONCLUSTERED (IsAlone),
 	INDEX IX_IsAStar NONCLUSTERED (IsAStar),
